backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
django-filter==23.5
djoser==2.1.0
gunicorn==20.1.0
webcolors==1.11.1
psycopg2-binary==2.9.3
djangorestframework-simplejwt==4.7.2
Pillow==9.3.0
PyJWT==2.1.0
requests==2.26.0
djoser==2.1.0
PyYAML==6.0
load_dotenv==0.1.0
django-urlshortner==0.0.2